Remove using service Endpoints for maintainOp
In Enter/Exit Maintenance Mode, we make two calls: `maintenanceOp` for
putting the node in maintenance mode, and `getPxNode` for checking the
status of the node.

The issue with `maintenanceOp` with using serviceEndpoint was that it
calls to the services and return a random node endpoint the service choose.
Therefore, putting a random node in maintenance and failing the status check.

This PR aims to resolve the issue. For now, we remove getting the service
endpoint.

Signed-off-by: dahuang <dahuang@purestorage.com>