PWX-23197: increase the timeout when waiting for the app destroy
We were using 3 min timeout while waiting for the app to get destroyed. This
was not enough in some cases when the app can legitimately take longer
to go away. Moreover, a few other tests are already using 5 min timeout
for destroy. Changed the code in k8s.go to use the 5 min timeout as well.