PWX-22448: add test-sharedv4 service app with aggr=2 and io_profile=db
Add another app to the sharedv4 function tests for the test cases
C55359 and C55360.

The test-sv4-svc-2 app uses aggregation_level=2, io_profile=db and repl=2.

A few tests that call getReplicaNodeIDs are not yet smart enough to deal
with the volumes with aggregation_level > 1. Changed those tests to
skip this new app.

Changed setHALevel function to not error out with the HA level is
already the same as the desired HA level.