PTX-12815: increase vol validation timeout when there are more volumes
We expect the PVCs to be ready within 2 min by default. We increase
that timeout when the app has several instances (app scale factor is
greater than 1). We need to do the same when an app is using
a lot of volumes.

This patch takes into account the number of volumes when setting
the timeout.

Signed-off-by: Neelesh Thakur <neelesh.thakur@purestorage.com>