PWX-41525: avoid picking a down node in GetAllKvdbNodes (#3437)
This was causing intermittent failures because we picked the node on which PX was down.

The problem was in GetAllKvdbNodes function which was modified by PR #3301 slightly incorrectly
because it should have picked the random node at randIndex from eligibleNodes but instead it
picked it from workerNodes potentially picking a down node. So, I initially thought that was
causing GetAllKvdbNodes to fail. But upon further inspection, we have one more call to
IsPxReadyOnNode. So, even if we initially picked a down random node, we should have skipped that
and proceeded to find another random node. But that did not happen which means that
IsPxReadyOnNode returned true for the node that we had just brought down. To fix this,
I have added an additional connectivity check in IsPxReadyOnNode. Hopefully, that should improve
detection of a down node in IsPxReadyOnNode.